Conflicts:
	apps/barneshut/Barneshut.cpp
	apps/delaunayrefinement/Mesh.h
	exp/apps/triangles/Triangles.cpp
	include/Galois/Galois.h
	include/Galois/Mem.h
	include/Galois/Runtime/Executor_DoAll.h
	include/Galois/Runtime/ParallelWork.h
	src/Support.cpp

+++ src/Support.cpp
@@ -104,7 +104,7 @@ public:
       auto& values = ii->second;
       size_t accum = 0;
       for (auto& x : values)
-        accum += std::accumulate(x.begin(), x.end(), 0);
+        accum += std::accumulate(x.begin(), x.end(), static_cast<unsigned long>(0));
 
       LL::gPrint("STAT,",
                  ii->first.first, ",",
